Move Config::from_args() to main.rs
